requirements.txt
# API Integration
openai>=1.0.0
google-api-python-client>=2.100.0

# Web Scraping and Processing
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3
html2text>=2020.1.16

# Database
pymongo>=4.5.0
motor>=3.3.0  # Async MongoDB driver

# Data Processing
pandas>=2.0.3
numpy>=1.24.0
scikit-learn>=1.3.0

# Rate Limiting and Backoff
ratelimit>=2.2.1
backoff>=2.2.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.3
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1

# API (if implementing REST API)
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2 